backend/requirements.txt
tinydb
fastapi
pydantic
uvicorn
python-dotenv
python-multipart